pyproject.toml
[project]
name = "story-telling-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-texttospeech>=2.27.0",
    "google-genai>=1.32.0",
    "google-generativeai>=0.8.5",
    "gtts>=2.5.4",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pyttsx3>=2.99",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "streamlit>=1.49.1",
    "streamlit-option-menu>=0.4.0",
]
